+++ src/Mzinga.Test/GameAITests.cs
@@ -2,11 +2,14 @@
 // Licensed under the MIT License.
 
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
+
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 using Mzinga.Core;
 using Mzinga.Core.AI;
+using Mzinga.Engine;
 
 namespace Mzinga.Test
 {
@@ -27,13 +30,15 @@ namespace Mzinga.Test
             TimeSpan sum = TimeSpan.Zero;
             int iterations = 100;
 
+            var config = EngineConfig.GetDefaultEngineConfig();
+
             for (int i = 0; i < iterations; i++)
             {
                 var gb = new Board();
-                GameAI ai = GetTestGameAI(gb.GameType);
+                GameAI ai = config.GetGameAI(gb.GameType);
 
                 Stopwatch sw = Stopwatch.StartNew();
-                _ = ai.GetBestMove(gb, 2, PerfTestMaxHelperThreads);
+                _ = ai.GetBestMove(gb, 2, config.MaxHelperThreads);
                 sw.Stop();
 
                 sum += sw.Elapsed;
@@ -49,15 +54,17 @@ namespace Mzinga.Test
             TimeSpan sum = TimeSpan.Zero;
             int iterations = 100;
 
+            var config = EngineConfig.GetDefaultEngineConfig();
+
             var original = GetBoardOnFifthTurn();
 
             for (int i = 0; i < iterations; i++)
             {
                 var gb = original.Clone();
-                GameAI ai = GetTestGameAI(gb.GameType);
+                GameAI ai = config.GetGameAI(gb.GameType);
 
                 Stopwatch sw = Stopwatch.StartNew();
-                _ = ai.GetBestMove(gb, 2, PerfTestMaxHelperThreads);
+                _ = ai.GetBestMove(gb, 2, config.MaxHelperThreads);
                 sw.Stop();
 
                 sum += sw.Elapsed;
@@ -70,8 +77,10 @@ namespace Mzinga.Test
         [TestCategory("Performance")]
         public void GameAI_FirstTurnBestMoveByTimePerfTest()
         {
+            var config = EngineConfig.GetDefaultEngineConfig();
+
             var gb = new Board();
-            GameAI ai = GetTestGameAI(gb.GameType);
+            GameAI ai = config.GetGameAI(gb.GameType);
 
             int maxDepth = 0;
 
@@ -80,7 +89,7 @@ namespace Mzinga.Test
                 maxDepth = Math.Max(maxDepth, e.Depth);
             };
 
-            Move m = ai.GetBestMove(gb, TimeSpan.FromSeconds(5), PerfTestMaxHelperThreads);
+            Move m = ai.GetBestMove(gb, TimeSpan.FromSeconds(5), config.MaxHelperThreads);
 
             Trace.WriteLine(string.Format("Max Depth: {0}", maxDepth));
         }
@@ -89,8 +98,10 @@ namespace Mzinga.Test
         [TestCategory("Performance")]
         public void GameAI_FifthTurnBestMoveByTimePerfTest()
         {
+            var config = EngineConfig.GetDefaultEngineConfig();
+
             var gb = GetBoardOnFifthTurn();
-            GameAI ai = GetTestGameAI(gb.GameType);
+            GameAI ai = config.GetGameAI(gb.GameType);
 
             int maxDepth = 0;
 
@@ -99,7 +110,7 @@ namespace Mzinga.Test
                 maxDepth = Math.Max(maxDepth, e.Depth);
             };
 
-            Move m = ai.GetBestMove(gb, TimeSpan.FromSeconds(5), PerfTestMaxHelperThreads);
+            Move m = ai.GetBestMove(gb, TimeSpan.FromSeconds(5), config.MaxHelperThreads);
 
             Trace.WriteLine(string.Format("Max Depth: {0}", maxDepth));
         }
@@ -108,11 +119,13 @@ namespace Mzinga.Test
         [TestCategory("Performance")]
         public void GameAI_DifferentHelperThreadsTest()
         {
+            var config = EngineConfig.GetDefaultEngineConfig();
+
             var gb = Board.ParseGameString(@"Base;InProgress;Black[17];wA1;bA1 wA1-;wA2 -wA1;bA2 bA1\;wQ \wA1;bQ bA2/;wA2 /bA2;bA3 \bQ;wA3 wA2\;bA3 wA3-;wG1 /wA2;bG1 bQ/;wQ -wA1;bA3 wA3\;wG1 bG1/;bS1 -bG1;wS1 /wA2;bS1 \wQ;wS1 bA3\;bG2 bG1\;wG2 -wA2;bS2 -bS1;wG2 bA2\;bG2 bS2\;wA2 \bS2;bB1 /bG2;wA2 -bS2;bB1 bG2\;wA2 wG2\;bS2 \wA1;wA2 \bA1;bB1 wQ\;wA2 bS2/");
 
             for (int maxHelpers = 0; maxHelpers < Environment.ProcessorCount / 2; maxHelpers++)
             {
-                GameAI ai = GetTestGameAI(gb.GameType);
+                GameAI ai = config.GetGameAI(gb.GameType);
                 Stopwatch sw = Stopwatch.StartNew();
                 var bestMove = ai.GetBestMove(gb, TimeSpan.FromSeconds(5), maxHelpers);
                 sw.Stop();
@@ -120,40 +133,55 @@ namespace Mzinga.Test
             }
         }
 
-        [TestMethod]
-        public void GameAI_CanSolveOneBestMoveToForceWinPuzzleTest()
+        [DataTestMethod]
+        [DynamicData(nameof(GetEngineConfigOptions), DynamicDataSourceType.Method)]
+        public void GameAI_CanSolveOneBestMoveToForceWinPuzzleTest(int maxHelperThreads, int quiescentSearchMaxDepth)
         {
-            TestUtils.LoadAndExecuteTestCases<GameAIBestMoveTestCase>("PuzzleCandidate_IsOneBestMoveToForceWinPuzzleTest.csv");
+            TestUtils.LoadAndExecuteTestCases<GameAIBestMoveTestCase>("PuzzleCandidate_IsOneBestMoveToForceWinPuzzleTest.csv", maxHelperThreads, quiescentSearchMaxDepth);
         }
 
-        [TestMethod]
-        public void GameAI_TreeStrapImprovesBoardScoreForOneBestMoveToForceWinPuzzleTest()
+        [DataTestMethod]
+        [DynamicData(nameof(GetEngineConfigOptions), DynamicDataSourceType.Method)]
+        public void GameAI_TreeStrapImprovesBoardScoreForOneBestMoveToForceWinPuzzleTest(int maxHelperThreads, int quiescentSearchMaxDepth)
         {
-            TestUtils.LoadAndExecuteTestCases<GameAITreeStrapTestCase>("PuzzleCandidate_IsOneBestMoveToForceWinPuzzleTest.csv");
+            TestUtils.LoadAndExecuteTestCases<GameAITreeStrapTestCase>("PuzzleCandidate_IsOneBestMoveToForceWinPuzzleTest.csv", maxHelperThreads, quiescentSearchMaxDepth);
         }
 
-        [TestMethod]
-        public void GameAI_BlockWinningMoveIsBestMoveTest()
+        [DataTestMethod]
+        [DynamicData(nameof(GetEngineConfigOptions), DynamicDataSourceType.Method)]
+        public void GameAI_BlockWinningMoveIsBestMoveTest(int maxHelperThreads, int quiescentSearchMaxDepth)
         {
-            TestUtils.LoadAndExecuteTestCases<GameAIBestMoveTestCase>("GameAI_BlockWinningMoveIsBestMoveTest.csv");
+            TestUtils.LoadAndExecuteTestCases<GameAIBestMoveTestCase>("GameAI_BlockWinningMoveIsBestMoveTest.csv", maxHelperThreads, quiescentSearchMaxDepth);
+        }
+
+        private static IEnumerable<object[]> GetEngineConfigOptions()
+        {
+            var maxHelperThreads = new int?[] {  EngineConfig.MinMaxHelperThreads, EngineConfig.MaxMaxHelperThreads };
+            var quiescentSearchMaxDepth = new int?[] { GameAIConfig.MinQuiescentSearchMaxDepth, GameAIConfig.MaxQuiescentSearchMaxDepth };
+
+            foreach (var mht in maxHelperThreads)
+            {
+                foreach (var qsmd in quiescentSearchMaxDepth)
+                {
+                    yield return new object[] { mht, qsmd };
+                }
+            }
         }
 
         private static Board GetBoardOnFifthTurn()
         {
             var gb = new Board();
-            GameAI ai = GetTestGameAI(gb.GameType);
+
+            var config = EngineConfig.GetDefaultEngineConfig();
+
+            GameAI ai = config.GetGameAI(gb.GameType);
 
             while (gb.CurrentPlayerTurn < 5)
             {
-                gb.TryPlayMove(ai.GetBestMove(gb, 1, TestMaxHelperThreads), "");
+                gb.TryPlayMove(ai.GetBestMove(gb, 1, config.MaxHelperThreads), "");
             }
 
             return gb;
-        }
-
-        public static GameAI GetTestGameAI(GameType gameType)
-        {
-            return TestUtils.DefaultGameEngineConfig.GetGameAI(gameType);
         }
 
         private class GameAIBestMoveTestCase : GameAIBestMoveTestCaseBase
@@ -162,8 +190,12 @@ namespace Mzinga.Test
 
             public override void Execute()
             {
-                GameAI ai = GetTestGameAI(Board.GameType);
-                ActualBestMove = ai.GetBestMove(Board, MaxDepth, TestMaxHelperThreads);
+                var config = EngineConfig.GetDefaultEngineConfig();
+                config._maxHelperThreads= TestArgs?[0] as int?;
+                config.QuiescentSearchMaxDepth = TestArgs?[1] as int?;
+
+                GameAI ai = config.GetGameAI(Board.GameType);
+                ActualBestMove = ai.GetBestMove(Board, MaxDepth, config.MaxHelperThreads);
                 Assert.AreEqual(ExpectedBestMove, ActualBestMove, $"Expected: {Board.GetMoveString(ExpectedBestMove)}, Actual: {Board.GetMoveString(ActualBestMove)}.");
             }
         }
@@ -172,7 +204,11 @@ namespace Mzinga.Test
         {
             public override void Execute()
             {
-                GameAI ai = GetTestGameAI(Board.GameType);
+                var config = EngineConfig.GetDefaultEngineConfig();
+                config._maxHelperThreads= TestArgs?[0] as int?;
+                config.QuiescentSearchMaxDepth = TestArgs?[1] as int?;
+
+                GameAI ai = config.GetGameAI(Board.GameType);
 
                 double colorValue = Board.CurrentColor == PlayerColor.White ? 1.0 : -1.0;
 
@@ -181,7 +217,7 @@ namespace Mzinga.Test
 
                 double startScore = colorValue * ai.CalculateBoardScore(Board);
 
-                ai.TreeStrap(Board, MaxDepth, TestMaxHelperThreads);
+                ai.TreeStrap(Board, MaxDepth, config.MaxHelperThreads);
                 ai.ResetCaches();
 
                 double endScore = colorValue * ai.CalculateBoardScore(clone);
@@ -192,6 +228,8 @@ namespace Mzinga.Test
 
         private abstract class GameAIBestMoveTestCaseBase : ITestCase
         {
+            public object[] TestArgs { get; set; }
+
             public Board Board;
             public int MaxDepth;
 
@@ -215,8 +253,5 @@ namespace Mzinga.Test
                 Assert.IsTrue(Board.TryParseMove(vals[2], out ExpectedBestMove, out string _));
             }
         }
-
-        private static readonly int TestMaxHelperThreads = 0;
-        private static readonly int PerfTestMaxHelperThreads = 0;
     }
 }